Fixed bug in Rename mode
There was a problem that, if we select the rename mode to a gameObject and we were editing another gameObject at the same time, it was displaying two text fields. This has been corrected and only one can me modified at the same time. If you try to rename one GameObject while you have another one being editing, the old one is confirmed and closed and the new one is in editing mode