Task/Enemy Behaviour THO-462 (#175)
* able to play without animations

* a bit of defensive programming

* Enemy rotation fixed when looking at player & adjustments on AI Agent Component

- Has been fixed the rotation of the Enemy when looking at player.
- Has been modified AI Agent Component due to some unexpected errors while being on playMode since the 2nd time in same engine execution.
-  Has been disabled the animation part of HandleState until animation Component is fixed.

* WIP: patrolling & type of character & anim prints comments

* speed in controllers

* WIP: first approach to patrolling

* WIP stop and resume movement AIAgentComponent

(WIP) Developed but not tested the stop and resume movement method in order to not make a robotic movement on the enemies.

- Remains to do the ignore lookAt if it is needed when enemy is paused.

* First version of natural behaviour

Enemy now fronzen in the position when enemy attacks player.

Note: It is not updated for the speed, instead is it done by maintaining the same position. Even though we change the speed, if the agent is not recreate it, the speed would be the same.

* rotation while patrolling

* patrol finished

* logs commented and some more merges with vs3

* const added

* another const

---------

Co-authored-by: Marc López <marclop.se@gmail.com>