Task/Level switch-THO-604  (#196)
* Added SceneChanger

Add scene change script, it is going to be tested if its working properly

* SceneChangeScript

Added the scene change script that triggers a change to another scene.

* broken filter

fixed a broken filter

* Fixed how to use the API

I changed the api to just be used on the header only. And erased the isOneUse since it was not used.

* Advances of respawnTimer

spawnTimer.

* Revert "Advances of respawnTimer"

This reverts commit 6db0baef6020d997b8b31ec3f96606e18a7883c8.