filter to ignore scalation and rotation
There were in some parts that the matris was counting the scalartion and rotation of the gameObject and was making unexpected behaviours in gizmo and internal calculations of world position for each Spline Point.

Now it is fixed and ignore the rotation and scalation part of the transform of the GameObject.