Task/Variable respawn points-THO-602  (#233)
* Added SceneChanger

Add scene change script, it is going to be tested if its working properly

* SceneChangeScript

Added the scene change script that triggers a change to another scene.

* broken filter

fixed a broken filter

* Fixed how to use the API

I changed the api to just be used on the header only. And erased the isOneUse since it was not used.

* Advances of respawnTimer

spawnTimer.

* Revert "Advances of respawnTimer"

This reverts commit 6db0baef6020d997b8b31ec3f96606e18a7883c8.

* Added RespawnController

added the respawn controller and methods in the character class to access certain properties.

* Finished Script

* Fixed issue bug

The health of healthpoints had a bug that assigned the value that didnt correspond when respawning

* Modfied

Modified the script a little to use the most recent vs3 changes

* Fixed merge errors

---------

Co-authored-by: Lukas <lukasganter@gmail.com>