Task/THO-567 smooth trails (#243)
* Basic skeleton coding for Spline Component and added in component Utils

Added the configuration, in order to the inpector detect it and be able to use it (not right now at the moment).

Also, created a base skeleton coding for spline component.

* basic add insert and remove point implementation

* Load constructor

* clone implemented

* See Spline on Inspector

* Added input float3 in inspector

* Temporal inspector Spline properties

Has been developed a temporal imgui section for spline component until is developed the core of Catmull-Rom formula calculation.

Some parts may change depending on the requirement that may be found in the future.

* Update format

* load and save loop conditional

* Implemented core formula of Centripetal Catmull-Rom

* Evaluation and evaluation segment implemented

* First version of Spline shown on scene

Now splines can be seen in the scene. Implemented DebugDraw for that.

Also has been updated DebugDrawModule to have the enable/disable view of splines. (At the moment is doing nothing. It is needed to take a look and solve this)

* Added drawing lines and curves without loop and having minimun 2 points

* Added getter world splinePoint position

* (WIP) Drawing gizmos in splinePoints

* update Spline points when moving GameObject parent

Now, when the GameObject owner of the Splice Component moves, points update correctly according to it.

(Gizmos have been disable in order to check correctly this part)

* (WIP) Gizmos

* Delete button for Spline Point

* Gizmos working in spline points

Now gizmos updates position of spline points. However, in order to do that, gameObject gizmos is disabled to be able to achieve that.

* Loop the entire spline with world position with t between 0-1

Loop the entire spline with a t value between 0 and 1 and also get its respective world position.

(with 3 points, the movement is not curved, looking into it to achieve that result)

* Updated Marker velocity dragFloat Imgui

* Travel marker updated (now accepts 2 or 3 points with curves)

Nor the travel marker travels along the curve with 3 points instead of going in a straight line between points.

* Include header spline component

* Added const & unnecessary continuous calculation

* updated unnecessary extra calculations

* filter to ignore scalation and rotation

There were in some parts that the matris was counting the scalartion and rotation of the gameObject and was making unexpected behaviours in gizmo and internal calculations of world position for each Spline Point.

Now it is fixed and ignore the rotation and scalation part of the transform of the GameObject.

* added const

* more const

* Trail base

* Trail component base

* A trail

* Bezier Curves

* ImGui color gradient

* imgui out

* ImGui curves moved to .cpp

* Trails WIdth and gradient

* sAVE COLOR

* Added texture imgui

* Refactor

* Smooth Trails

* Changed

---------

Co-authored-by: Marc López <marclop.se@gmail.com>
Co-authored-by: Marc López <88272783+MarcLopezS@users.noreply.github.com>