Task/spline game object rotation tho 663 (#256)
* Added SplinePoint struct & Save/Load method updated

updated save and load method, in order to accept struct SplinePoint.

* (WIP) implemented EvaluateRotation

* ImGui Rotation parameters

* Cone marker to see rotation

* added Gizmo Rotation in SplinePoints & cones renderDebug

Now the spline points are represented as a cone instead of a sphere in order to see which rotation would have the gameObject that would route the Spline.

Also, gizmos rotation has been added to splinePoints.

* Adapted Rotation to Script MoveGOInSpline

Rotation is now considered in the script example of moving a GameObject through Spline.

* Added speed variable in spline points & save/load speed

* Interpolating speed

(Note: Thinking in the future make a globalization with Evaluate, EvaluateRotation and EvaluateSpeed due to almost have the exact same code)

* Speed controller in MoveGOInSpline

Now the script takes the interpolated speed of the splinePoint and applies it to the movement. Taking into account that calculates the variation of the speed between two points, depending on which part of the segment of the spline.