Player progression (lock & unlock skills) final Player THO-777 (#329)
* fix Dash UI Cooldown for progression (only UI)

* update spawnUI & added abilityUnlocker (WIP)

* Unlock ability functionality

* Limit unlocking dash

Now dash can only be activated only if player has crossed the dash tutorial.

* added unlocked checkbox player ability from start

This makes sure which abilities are activated in the scene and which not.

There is a problem in which the checkboxes are not saved when saving the project. Needs to be looked and make it work so there are no unexpected behaviour.

* move unlock inspector fields to avoid inspector bug

* updated name ultimate condition