added temp player name
Due to the change scene, PlayerName value is lost. In order to keep the name temporaly, we make a copy of it.

Now the scene works everything except the riastrad bar that, even that the value is loaded correctly, the riastrad bar is not updated until it increases again. Also, it is needed to save the number of mushrooms picked.