Task/player save file progression (#354)
* Fix proyectile bug.

* Getters & setters for health.

* Getters & Setters for CuChulainn

* Added auxiliary file to allocate save/load player data functions

* first version of Save Player data system defined

* New methods to export and save state in Cuchulainn

* Added save data when changing scene and various fixes

* added temp player name

Due to the change scene, PlayerName value is lost. In order to keep the name temporaly, we make a copy of it.

Now the scene works everything except the riastrad bar that, even that the value is loaded correctly, the riastrad bar is not updated until it increases again. Also, it is needed to save the number of mushrooms picked.

* added mushrooms

* fixed update riastrad UI on changing scene

* Added game Session

Game session is a helper when the player starts a new game, if a save file exist, ignore it and set the player values by default. Can be adapted to a continue function if it is wanted.

* Delete save file when starting new game

* removed save & load from pressAnyKey script

Save & load method have been removed because they were overriding those methods from script component and was causing unexpected behaviour like not saving the inputs given. Now it is fixed.